Source290513-2300Uhr plus Airplane

+++ src/drv_hmc5883l.c
@@ -9,11 +9,11 @@
 #define HMC58X3_R_CONFA 0
 #define HMC58X3_R_CONFB 1
 #define HMC58X3_R_MODE 2
-#define HMC58X3_X_SELF_TEST_GAUSS (+1.16)       // X axis level when bias current is applied.
-#define HMC58X3_Y_SELF_TEST_GAUSS (+1.16)       // Y axis level when bias current is applied.
-#define HMC58X3_Z_SELF_TEST_GAUSS (+1.08)       // Y axis level when bias current is applied.
-#define SELF_TEST_LOW_LIMIT  (243.0 / 390.0)    // Low limit when gain is 5.
-#define SELF_TEST_HIGH_LIMIT (575.0 / 390.0)    // High limit when gain is 5.
+#define HMC58X3_X_SELF_TEST_GAUSS (+1.16f)       // X axis level when bias current is applied.
+#define HMC58X3_Y_SELF_TEST_GAUSS (+1.16f)       // Y axis level when bias current is applied.
+#define HMC58X3_Z_SELF_TEST_GAUSS (+1.08f)       // Y axis level when bias current is applied.
+#define SELF_TEST_LOW_LIMIT  (243.0f / 390.0f)   // Low limit when gain is 5.
+#define SELF_TEST_HIGH_LIMIT (575.0f / 390.0f)   // High limit when gain is 5.
 #define HMC_POS_BIAS 1
 #define HMC_NEG_BIAS 2
 
@@ -35,64 +35,64 @@ void hmc5883lInit(float *calibrationGain)
     float magGain[3];
     int16_t magADC[3];
     int i;
-    int32_t xyz_total[3] = { 0, 0, 0 }; // 32 bit totals so they won't overflow.
-    bool bret = true;           // Error indicator
+    int32_t xyz_total[3] = { 0, 0, 0 };             // 32 bit totals so they won't overflow.
+    bool bret = true;                               // Error indicator
 
-    // PB12 - MAG_DRDY output on rev4 hardware
-    GPIO_InitStructure.GPIO_Pin = GPIO_Pin_12;
+    GPIO_InitStructure.GPIO_Pin = GPIO_Pin_12;      // PB12 - MAG_DRDY output on rev4 hardware
     GPIO_InitStructure.GPIO_Speed = GPIO_Speed_2MHz;
     GPIO_InitStructure.GPIO_Mode = GPIO_Mode_IN_FLOATING;
     GPIO_Init(GPIOB, &GPIO_InitStructure);
 
     delay(50);
     i2cWrite(MAG_ADDRESS, HMC58X3_R_CONFA, 0x010 + HMC_POS_BIAS);   // Reg A DOR = 0x010 + MS1, MS0 set to pos bias
-    // Note that the  very first measurement after a gain change maintains the same gain as the previous setting. 
+    // Note that the  very first measurement after a gain change maintains the same gain as the previous setting.
     // The new gain setting is effective from the second measurement and on.
     i2cWrite(MAG_ADDRESS, HMC58X3_R_CONFB, 2 << 5); // Set the Gain
     delay(100);
     hmc5883lRead(magADC);
 
-    for (i = 0; i < 10; i++) {  // Collect 10 samples
+    for (i = 0; i < 10; i++)                        // Collect 10 samples
+    {
         i2cWrite(MAG_ADDRESS, HMC58X3_R_MODE, 1);
         delay(50);
-        hmc5883lRead(magADC);       // Get the raw values in case the scales have already been changed.
+        hmc5883lRead(magADC);                       // Get the raw values in case the scales have already been changed.
 
-        // Since the measurements are noisy, they should be averaged rather than taking the max.
-        xyz_total[0] += magADC[0];
+        xyz_total[0] += magADC[0];                  // Since the measurements are noisy, they should be averaged rather than taking the max.
         xyz_total[1] += magADC[1];
         xyz_total[2] += magADC[2];
 
-        // Detect saturation.
-        if (-4096 >= min(magADC[0], min(magADC[1], magADC[2]))) {
+        if (-4096 >= min(magADC[0], min(magADC[1], magADC[2])))     // Detect saturation.
+        {
             bret = false;
-            break;              // Breaks out of the for loop.  No sense in continuing if we saturated.
+            break;                                  // Breaks out of the for loop.  No sense in continuing if we saturated.
         }
         LED1_TOGGLE;
     }
 
     // Apply the negative bias. (Same gain)
     i2cWrite(MAG_ADDRESS, HMC58X3_R_CONFA, 0x010 + HMC_NEG_BIAS);   // Reg A DOR = 0x010 + MS1, MS0 set to negative bias.
-    for (i = 0; i < 10; i++) {
+    for (i = 0; i < 10; i++)
+    {
         i2cWrite(MAG_ADDRESS, HMC58X3_R_MODE, 1);
         delay(50);
-        hmc5883lRead(magADC);               // Get the raw values in case the scales have already been changed.
+        hmc5883lRead(magADC);                       // Get the raw values in case the scales have already been changed.
 
-        // Since the measurements are noisy, they should be averaged.
-        xyz_total[0] -= magADC[0];
+        xyz_total[0] -= magADC[0];                  // Since the measurements are noisy, they should be averaged.
         xyz_total[1] -= magADC[1];
         xyz_total[2] -= magADC[2];
 
         // Detect saturation.
-        if (-4096 >= min(magADC[0], min(magADC[1], magADC[2]))) {
+        if (-4096 >= min(magADC[0], min(magADC[1], magADC[2])))
+        {
             bret = false;
-            break;              // Breaks out of the for loop.  No sense in continuing if we saturated.
+            break;                                  // Breaks out of the for loop.  No sense in continuing if we saturated.
         }
         LED1_TOGGLE;
     }
 
-    magGain[0] = fabs(820.0 * HMC58X3_X_SELF_TEST_GAUSS * 2.0 * 10.0 / xyz_total[0]);
-    magGain[1] = fabs(820.0 * HMC58X3_Y_SELF_TEST_GAUSS * 2.0 * 10.0 / xyz_total[1]);
-    magGain[2] = fabs(820.0 * HMC58X3_Z_SELF_TEST_GAUSS * 2.0 * 10.0 / xyz_total[2]);
+    magGain[0] = fabs(820.0 * HMC58X3_X_SELF_TEST_GAUSS * 2.0f * 10.0f / (float)xyz_total[0]);
+    magGain[1] = fabs(820.0 * HMC58X3_Y_SELF_TEST_GAUSS * 2.0f * 10.0f / (float)xyz_total[1]);
+    magGain[2] = fabs(820.0 * HMC58X3_Z_SELF_TEST_GAUSS * 2.0f * 10.0f / (float)xyz_total[2]);
 
     // leave test mode
     i2cWrite(MAG_ADDRESS, HMC58X3_R_CONFA, 0x70);   // Configuration Register A  -- 0 11 100 00  num samples: 8 ; output rate: 15Hz ; normal measurement mode
@@ -100,14 +100,15 @@ void hmc5883lInit(float *calibrationGain)
     i2cWrite(MAG_ADDRESS, HMC58X3_R_MODE, 0x00);    // Mode register             -- 000000 00    continuous Conversion Mode
     delay(100);
 
-    if (!bret) {                // Something went wrong so get a best guess
-        magGain[0] = 1.0;
-        magGain[1] = 1.0;
-        magGain[2] = 1.0;
+    if (!bret)                                      // Something went wrong so get a best guess
+    {
+        magGain[0] = 1.0f;
+        magGain[1] = 1.0f;
+        magGain[2] = 1.0f;
     }
 
-    // if parameter was passed, give calibration values back
-    if (calibrationGain) {
+    if (calibrationGain)                            // if parameter was passed, give calibration values back
+    {
         calibrationGain[0] = magGain[0];
         calibrationGain[1] = magGain[1];
         calibrationGain[2] = magGain[2];
